% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{bib_check}
\alias{bib_check}
\title{Check Whether the Login Page Is Reachable}
\usage{
bib_check(silent = TRUE)
}
\arguments{
\item{silent}{logical: should error messages be suppressed?}
}
\value{
logical indicating whether the login page could be reached
}
\description{
Check whether the login page can be reached. If this fails, either there is
no internet connection or the server of the library is down.
}
