% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_alert.R
\name{bib_email_alert}
\alias{bib_email_alert}
\title{Send Email Alert if Rental Runs Out For Some Documents}
\usage{
bib_email_alert(
  users,
  n_days,
  recipients,
  host,
  username,
  password,
  from = username,
  port = 465,
  verbose = FALSE
)
}
\arguments{
\item{users}{a list of users with username and password.
Such a list can be created from a JSON file using
\code{\link{bib_read_login_data}}.}

\item{n_days}{numeric. Rentals that run out within \code{n_days} days from
today will cause an alert message to be sent.}

\item{recipients}{e-mail addresses of one or several recipients.}

\item{host}{DNS name or IP address of the SMTP server.}

\item{username, password}{username and password for the SMTP server.}

\item{from}{email address to use as the sender, by default equal to
\code{username}. In principle, you can use an arbitrary address here, but some
email providers reject emails where the domain of the sender does not
match the host.}

\item{port}{Port that the SMTP server is listening on.}

\item{verbose}{should verbose output be produced from the interaction
with the SMTP server?}
}
\value{
\code{TRUE} if the function was successful, i.e., if either an alert was sent
successfully or no alert was needed because no rentals are running out within
the given time frame. If any error occurred (e.g., during login or when
sending the email), \code{FALSE} is returned.
}
\description{
Send an email alert if the rental of some documents runs out
within a given number of days.
}
