% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{bib_encrypt}
\alias{bib_encrypt}
\title{Encrypt Password}
\usage{
bib_encrypt(password)
}
\arguments{
\item{password}{character giving the password to be encrypted}
}
\description{
The page uses simple encryption to make the transfer of passwords more
secure. This function applies the appropriate encryption to generate a
password that can be used for login.
}
