% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_export.R
\name{bib_excel_method}
\alias{bib_excel_method}
\title{Check whether Excel Export is Possible}
\usage{
bib_excel_method()
}
\value{
a character string with one of the values \code{"WriteXLS"}, \code{"writexl"} or
\code{"none"} indicating the method that will be used for Excel export.
}
\description{
Check which, if any, of the two variants of Excel export supported by
\code{bibleRe} are currently available on the system.
}
\details{
\code{bibleRe} supports Excel export with two packages:
\describe{
\item{\code{WriteXLS}}{This is the prefered method, but it requires a
perl installation. Use \code{\link[WriteXLS:testPerl]{WriteXLS::testPerl()}} to check whether perl
and all the required modules are installed. This writes formatted
Excel files with fixed first row and adjusted column width.}
\item{\code{writexl}}{This is used, when \code{WriteXLS} is not installed and has
the advantage that it has no external dependencies. However, the Excel files
do not have a fixed first row and the columns all have default width.}
}

The function checks the availability of each method and returns the name
of the package that will be used for writing Excel files with
\code{\link[=bib_write_excel]{bib_write_excel()}}.
}
