% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{bib_get_icon}
\alias{bib_get_icon}
\title{Download the Favicon from the Library's Website}
\usage{
bib_get_icon(file = "~/biblere.ico")
}
\arguments{
\item{file}{path where the icon will be stored. If the name does
not end with ".ico", the ending is appended.}
}
\description{
Download the Favicon from the Library's Website
}
\details{
This requires the package \code{faviconPlease} to be installed. Install it
with \code{install.packages('faviconPlease')}.
}
