% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{bib_list_documents}
\alias{bib_list_documents}
\title{List Borrowed Documents}
\usage{
bib_list_documents(session)
}
\arguments{
\item{session}{a logged in session to the library
web interface}
}
\value{
a tibble containing a table of borrowed documents
}
\description{
List Borrowed Documents
}
