% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fees.R
\name{bib_list_fees}
\alias{bib_list_fees}
\title{List Fees}
\usage{
bib_list_fees(session)
}
\arguments{
\item{session}{a logged in session to the library
web interface}
}
\value{
a tibble containing a table of fees
}
\description{
List Fees
}
