% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{bib_list_orders}
\alias{bib_list_orders}
\title{List Orders}
\usage{
bib_list_orders(session)
}
\arguments{
\item{session}{a logged in session to the library
web interface}
}
\value{
a tible containing a table of orders
}
\description{
List Orders
}
