% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchlist.R
\name{bib_list_watchlist}
\alias{bib_list_watchlist}
\title{Download Watchlist}
\usage{
bib_list_watchlist(session)
}
\arguments{
\item{session}{a logged in session to the library
web interface}
}
\value{
a tibble containing a table items on the watchlist
}
\description{
Download a table of items on a users watchlist.
}
\seealso{
Other functions to download data: 
\code{\link{bib_list_documents}()},
\code{\link{bib_list_fees}()},
\code{\link{bib_list_orders}()}
}
\concept{functions to download data}
