% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{bib_login}
\alias{bib_login}
\title{Login to the Web Interface}
\usage{
bib_login(username, password, displayname = NULL)
}
\arguments{
\item{username}{character giving the username, which usually
consists of 6 digits.}

\item{password}{character giving the password. See also
the security warning under 'Details'.}

\item{displayname}{optional character giving the name of the user to be used
in log messages. If omitted, the log messages show \code{username}.}
}
\value{
a \code{session} object
}
\description{
Login to the web interface with username and password.
}
\details{
\strong{Security warning:}
It is possible that passwords are stored in clear text and
can even be seen by employees of the library. This has
been the case at some point and it is not known whether
this has changed in the meantime. Your email address is stored
as well. Do not use a password that is also used anywhere else!
}
