% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{bib_logout}
\alias{bib_logout}
\title{Log out a running session}
\usage{
bib_logout(session)
}
\arguments{
\item{session}{the session object of the session with the web interface.}
}
\value{
x invisibly. The \code{session} is modified as a side effect.
}
\description{
Log out from a running session with the web interface.
}
