% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{bib_read_login_data}
\alias{bib_read_login_data}
\title{Read File With Login Data}
\usage{
bib_read_login_data(file)
}
\arguments{
\item{file}{character giving the name of a file.}
}
\description{
Read a JSON file with login data for one or more
users.
}
\details{
The JSON-file must have the following format:\if{html}{\out{<div class="sourceCode json">}}\preformatted{\{
  "User1": \{
    "username": "123456",
    "password": "password1"
  \},
  "User2": \{
    "username": "104392",
    "password": "password2"
  \},
  "User3": \{
    "username": "101010",
    "password": "password3"
  \}
\}
}\if{html}{\out{</div>}}

A block must be included for each user and must contain
the fields \code{username} and \code{password}. The above example is
also contained as a file in the package. You can obtain
the path to the file with\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("example/biblere_passwords.json",
            package = "bibleRe")
}\if{html}{\out{</div>}}
}
