% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew.R
\name{bib_renew}
\alias{bib_renew}
\title{Renew Documents}
\usage{
bib_renew(session, chk_ids)
}
\arguments{
\item{session}{a logged in session to the library
web interface}

\item{chk_ids}{character giving the checkbox ids of the
documents to be extended. Note that those documents
should belong to the user that is currently logged in.}
}
\description{
Renew Documents
}
