% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{bib_search}
\alias{bib_search}
\title{Search the Library Database}
\usage{
bib_search(search)
}
\arguments{
\item{search}{a character giving the search string}
}
\description{
Search the Library Database
}
\details{
\code{bib_search()} returns the link to be used for the search
as character string.

\code{bib_run_search()} uses javascript to open the search link
in a new tab.
}
