% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_export.R
\name{bib_setup_excel_export}
\alias{bib_setup_excel_export}
\title{Install packages for Excel export}
\usage{
bib_setup_excel_export(pkgs = c("WriteXLS", "writexl"))
}
\arguments{
\item{pkgs}{character vector giving the packages to be installed. By default,
both packages are tried, but \code{WriteXLS} is only installed if a \code{perl}
installation is found.}
}
\value{
character vector with the names of the packages that can be used for
Excel export.
}
\description{
Excel export with \code{\link{bib_write_excel}} uses either \code{\link[WriteXLS:WriteXLS]{WriteXLS::WriteXLS()}}
(prefered) or \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}} to create Excel files. This function
installs the required packages. \code{WriteXLS} is only installed, if a perl
installation is found.
}
\details{
Packages that are already installed won't be reinstalled.
}
