% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_export.R
\name{bib_write_excel}
\alias{bib_write_excel}
\title{Write a Table to an Excel File}
\usage{
bib_write_excel(
  table,
  file,
  type = c("documents", "orders", "fees", "watchlist")
)
}
\arguments{
\item{table}{a tibble containing documents, orders, or fees as returned
by the functions \code{\link[=bib_list_documents]{bib_list_documents()}}, \code{\link[=bib_list_fees]{bib_list_fees()}},
\code{\link[=bib_list_orders]{bib_list_orders()}}, and \code{\link[=bib_get_all_data]{bib_get_all_data()}}. The table may be
filtered, but columns should be left intact.}

\item{file}{path to a file}

\item{type}{character indicating the type of the table to be written. This
is required to translate column names to German and to set appropriate
sheet names in the Excel file.}
}
\value{
logical indicating write success. An Excel file is create as a side effect.
}
\description{
Write the table of documents, orders, or fees to an Excel file.
}
\details{
The method used for writing is determined by the availability of the
packages \code{WriteXLS} and \code{writexl}. Which method will be used can be checked
using \code{\link[=bib_excel_method]{bib_excel_method()}}.
}
