% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibleRe-package.R
\docType{package}
\name{bibleRe-package}
\alias{bibleRe}
\alias{bibleRe-package}
\title{Access Web Interface of the Public Library in Koeniz}
\description{
The public library of Koeniz, Switzerland, offers a web interface that allows users to check their borrowed documents, renew the rental period, etc. This package offers an interface to use some of that functionality. It is particularly useful to manage multiple accounts simultaneously.
}
\keyword{internal}
