% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{flex_html_table}
\alias{flex_html_table}
\title{Parse Potentially Faulty HTML Table}
\usage{
flex_html_table(tab_node)
}
\arguments{
\item{tab_node}{a node set containing a table}
}
\description{
Some html tables on the library site are faulty because they use \verb{<td>}
instead of \verb{<th>} in the table header. As a consequence, \code{rvest::html_table()}
does not use the headers as column names. This function recognises that
mistake and fixes the column headers.

The function also removes columns without name and replaces HTML line
breaks (\verb{<br>}) by \verb{\\n}.
}
