% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{bib_check}
\alias{bib_check}
\title{Check that Login Page is Reachable}
\usage{
bib_check()
}
\value{
logical indicating whether the login page could be reached
}
\description{
Check that the login page can be reached. If this fails, either there is
no internet connection or the server of library is down.
}
