% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_biblere}
\alias{run_biblere}
\title{Run the application}
\usage{
run_biblere(
  login_data_file = "~/.biblere_passwords",
  n_due_days = 7,
  launch.browser = NULL
)
}
\arguments{
\item{login_data_file}{character giving the full path to
a file containing the login data.}

\item{n_due_days}{integer giving the default for the selection
of the date used to filter due dates. By default, the app will
show documents that are due at most \code{n_due_days} days form today.}

\item{launch.browser}{logical, if \code{TRUE}, the application
is opened in the system's default browser, if \code{FALSE},
no browser is started. If the argument is omitted, the value
according to the option \code{shiny.launch.browser} is used,
which in RStudio opens the internal shiny viewer.}
}
\description{
Run the application
}
